Create map w/UIViewRepresentable to get coordinate w/longPressGestureRecognizer
