+++ SwiftBuddiesIOS/Targets/MapModule/Sources/Views/NewEventView.swift
@@ -5,6 +5,7 @@
 //  Created by Oğuzhan Abuhanoğlu on 13.05.2024.
 //
 
+import Foundation
 import SwiftUI
 import MapKit
 import SwiftData
@@ -13,11 +14,8 @@ struct NewEventView: View {
     
     @Environment(\.presentationMode) var presentationMode
     @Environment(\.modelContext) private var context
-    @Query private var items: [EventModel]
     
     @StateObject var vm = MapViewModel()
-
-    //@State private var tappedLocation: CLLocationCoordinate2D?
     
     private let categories = [
         "Meeting",
@@ -33,7 +31,7 @@ struct NewEventView: View {
     @State var adressText: String = ""
     @State var startDate: Date = Date()
     @State var dueDate: Date = Date()
-    //@State var location: CLLocationCoordinate2D =
+    @State var tappedLocation: CLLocationCoordinate2D?
    
 
     var body: some View {
@@ -59,6 +57,9 @@ struct NewEventView: View {
        
         
     }
+    
+    
+    
     
 }
 
@@ -157,10 +158,9 @@ extension NewEventView {
     }
     
     private var mapLayer: some View {
+        
         VStack {
-            Map(position: $vm.position) {
-                
-            }
+            MapViewRepresentable(tappedLocation: $tappedLocation)
             .mapControls {
                 MapUserLocationButton()
                 MapPitchToggle()
@@ -168,19 +168,21 @@ extension NewEventView {
             .onAppear{
                 CLLocationManager().requestWhenInUseAuthorization()
             }
-
         }
         .aspectRatio(1, contentMode: .fill)
         .cornerRadius(15)
         .padding(.horizontal)
+         
     }
+    
+    
     
     private var createButton: some View {
         Button(action: {
             // Save the event into core data
             if selectedCategory != nil {
                 vm.addItem(modelContext: context, id: UUID().uuidString, category: selectedCategory ?? "", name: nameText, about: descriptionText, startDate: startDate, dueDate: dueDate)
-                print(items.count)
+                print(tappedLocation?.latitude)
                 self.presentationMode.wrappedValue.dismiss()
             } else {
               //Hata mesajı
@@ -200,71 +202,3 @@ extension NewEventView {
 
 
 
-/*import SwiftUI
-import MapKit
-
-struct MapView: UIViewRepresentable {
-    @Binding var region: MKCoordinateRegion
-    @Binding var annotations: [MKPointAnnotation]
-    
-    class Coordinator: NSObject, MKMapViewDelegate {
-        var parent: MapView
-        
-        init(_ parent: MapView) {
-            self.parent = parent
-        }
-        
-        @objc func handleLongPress(gestureRecognizer: UILongPressGestureRecognizer) {
-            if gestureRecognizer.state == .began {
-                let location = gestureRecognizer.location(in: gestureRecognizer.view as? MKMapView)
-                if let mapView = gestureRecognizer.view as? MKMapView {
-                    let coordinate = mapView.convert(location, toCoordinateFrom: mapView)
-                    let annotation = MKPointAnnotation()
-                    annotation.coordinate = coordinate
-                    parent.annotations.append(annotation)
-                }
-            }
-        }
-    }
-    
-    func makeCoordinator() -> Coordinator {
-        return Coordinator(self)
-    }
-    
-    func makeUIView(context: Context) -> MKMapView {
-        let mapView = MKMapView()
-        mapView.delegate = context.coordinator
-        
-        let longPressGesture = UILongPressGestureRecognizer(target: context.coordinator, action: #selector(Coordinator.handleLongPress(gestureRecognizer:)))
-        mapView.addGestureRecognizer(longPressGesture)
-        
-        return mapView
-    }
-    
-    func updateUIView(_ uiView: MKMapView, context: Context) {
-        uiView.setRegion(region, animated: true)
-        uiView.removeAnnotations(uiView.annotations)
-        uiView.addAnnotations(annotations)
-    }
-}
-
-struct ContentView: View {
-    @State private var region = MKCoordinateRegion(
-        center: CLLocationCoordinate2D(latitude: 37.7749, longitude: -122.4194),
-        span: MKCoordinateSpan(latitudeDelta: 0.05, longitudeDelta: 0.05)
-    )
-    @State private var annotations = [MKPointAnnotation]()
-    
-    var body: some View {
-        VStack {
-            MapView(region: $region, annotations: $annotations)
-                .aspectRatio(1, contentMode: .fill)
-                .cornerRadius(15)
-                .padding(.horizontal)
-        }
-        .onAppear {
-            CLLocationManager().requestWhenInUseAuthorization()
-        }
-    }
-}
-*/
